Improve directory structure bit
