+++ modular-particle-system/src/serialization/moduleSerialization.ts
